requirements.txt
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
cloudscraper==1.2.71
contourpy==1.3.0
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
Jinja2==3.1.4
kiwisolver==1.4.7
lxml==5.3.0
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
playwright==1.46.0
pyee==11.1.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
requests-toolbelt==1.0.0
selenium==4.24.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
